Error handler moved to dedicated file.
