+++ src/index.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { StrictMode, useEffect, useState } from 'react';
 import { createRoot } from 'react-dom/client';
 import { HashRouter, Routes, Route, Navigate } from 'react-router-dom';
@@ -12,41 +11,7 @@ import { DnsCheckHome, DnsCheckResult } from './pages/dnschecker';
 import StringConversionPage from './pages/converter';
 import { CronConversionPage } from './pages/cron';
 import UnixEpochPage from './pages/time';
-
-interface ErrorProp {
-	children: any;
-}
-
-interface StateProp {
-	hasError: boolean;
-}
-
-class ErrorBoundary extends React.Component<ErrorProp, StateProp> {
-	constructor(props:any) {
-		super(props);
-		this.state = { hasError: false };
-	}
-
-	static getDerivedStateFromError(error:any) {
-		return { hasError: true };
-	}
-
-	render() {
-		if (this.state.hasError) {
-			return (
-				<div style={{fontFamily: 'sans-serif', textAlign: 'center'}}>
-					<p style={{fontSize: '8em', margin: 0}}>:(</p>
-					<h1>A crash has occurred</h1>
-					<p>We apologise for the inconvinience. Click the button below to return.</p>
-					<button onClick={() => (window.location.reload())}>Reload</button>
-				</div>
-			);
-		}
-
-		return this.props.children;
-	}
-}
-
+import { ErrorBoundary } from "./error";
 
 export default function App() {
 	// https://stackoverflow.com/a/65049865
